+++ lib/cache/memory_test.go
@@ -141,4 +141,46 @@ func TestMemoryCacheCompaction(t *testing.T) {
 	}
 }
 
+func TestMemoryCacheInitValues(t *testing.T) {
+	conf := NewConfig()
+	conf.Type = "memory"
+	conf.Memory.TTL = 0
+	conf.Memory.CompactionInterval = ""
+	conf.Memory.InitValues = map[string]string{
+		"foo":  "bar",
+		"foo2": "bar2",
+	}
+
+	c, err := New(conf, nil, log.Noop(), metrics.Noop())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	exp := "bar"
+	if act, err := c.Get("foo"); err != nil {
+		t.Error(err)
+	} else if string(act) != exp {
+		t.Errorf("Wrong result: %v != %v", string(act), exp)
+	}
+
+	// This should trigger compaction.
+	if err = c.Add("foo3", []byte("bar3")); err != nil {
+		t.Error(err)
+	}
+
+	exp = "bar"
+	if act, err := c.Get("foo"); err != nil {
+		t.Error(err)
+	} else if string(act) != exp {
+		t.Errorf("Wrong result: %v != %v", string(act), exp)
+	}
+
+	exp = "bar2"
+	if act, err := c.Get("foo2"); err != nil {
+		t.Error(err)
+	} else if string(act) != exp {
+		t.Errorf("Wrong result: %v != %v", string(act), exp)
+	}
+}
+
 //------------------------------------------------------------------------------